[AWARE] Clarify semantics of service discovery parameters
On service discovery which happens on the subscriber we receive
some information about the peer (the publisher): it's id, some
arbitrary blob it's providing (ssi), and a "match filter". The
match filter is the matching criteria (aside from service name)
used to find a service. The semantics are clarified by this CL:

- Unsolicited publish/passive subscriber: the publisher is active
  and broadcasting. We're getting the match filter which it
  advertises. No change.

- Solicited publish/active subscriber: the publisher is silent
  and will only respond if it matches what the subscriber provides.
  In such a case we will never know it's exact criteria and
  therefore cannot provide it to the app. We're simply providing
  the local match criteria. This is a clarification/possible
  change from initial understanding.

Bug: 36371061
Test: DiscoveryTest:test_positive_solicited_active_typical
Change-Id: I7d989cf0dc2383d15e4df18c9c0ea079f0486d80