Remove android.test dependencies on a.o.PerformanceCollector
The a.t.TimedTest annotation was removed because it was not in the API
and was only used by the code that has been removed.

The a.t.PerformanceCollectorTestCase was removed because it was not in
the API and its only purpose was to allow the AndroidTestRunner to
supply a PerformanceCollector.PerformanceResultsWriter to a test.

Bug: 62652137
Test: make checkbuild
Change-Id: Ie90a7004ade1448cb6bc7e029f6ba5620db29bfd