Always run Virtual Display when in a BootsToVR device.
This CL:
1) Moved initialization of the virtual display adapter to happen along
side the default display.  On standalone VR devices, a virtual display
is used as the main display for 2D UI so it is necessary that the virtual
display system is ready after PHASE_WAIT_FOR_DEFAULT_DISPLAY.
2) Add knowledge of bootsToVr in Vr2dDisplay, indicating that the device
is a standalone device that boots into VR. This forces Vr2dDisplay to
create the virtualDisplay at PHASE_SYSTEM_SERVICES_READY and never
destroy it (so that the activity stack doesn't go away).
3) Add tests to ensure default display and virtual display startup
sequence doesn't change.

Change-Id: Ie748771abd0b115db3c65a66a4b7a98dd1b28f2e
Bug:62842592
Test: Start a polaris device for first time, ensure SUW runs with
Virtual display dimentions.
Test: DisplayManagerServiceTest.java