fix: android/multinetwork.h can't be imported when using libandroid_net
libandroid_net is a subset of libandroid exposing the multinetwork
functionality. Contrary to libandroid which is not available to vendors,
libandroid_net is available to vendors as ll-ndk.

However, the header path for the two libraries are different.

for libandroid:
include <android/multinetwork.h>

for libandroid_net:
include <multinetwork.h>

This is because multinetwork.h is located in
<home_of_libandroid_net>/include directory. Since many vendors are using
<android/multinetwork.h>, we need to have a subdirectory 'android'.

Bug: 63120916
Test: build
Change-Id: I1cbc6e1641edda2013c80f9b16b825f0854776c1