[P2P][AsyncChannel] Revert client registering for service death
This effectively reverts 2f1cebee3f7566bb864c7532d32e0c9a4da67a81.

The change added a binder death listener on a peer of the AsyncChannel.

Issue: as specified in the bug - this causes a memory leak in apps
using WifiP2pManager (and possibly other clients using AsyncChannel).
The issue is that a reference to the caller is now kept by the death
listener. This means that the caller object cannot be gc'd. This
impacts WifiP2pManager since it didn't have an explicit close() method.

The original reason for the change was to enable clean-up - this can
be done in an alternative mechanism by passing-in a binder from client
to service. Will be enable in a subsequent CL.

Bug: 37443149
Test: all p2p tests, memory usage with test app
Change-Id: Iff044a8cc9c2fc0bc57db4a218b80d0b12424ddd